refactor: refactor NeuroMC op_counter into modular package
