+++ spikingjelly/activation_based/op_counter/__init__.py
@@ -4,5 +4,4 @@ from .memory_access import *
 from .ac import *
 from .synop import *
 from .mac import *
-from .neuromc_counters import *
-from .neuromc_energy import *
+from .neuromc import *
