+++ spikingjelly/activation_based/triton_kernel/surrogate_kernel.py
@@ -1,4 +1,3 @@
-import math
 
 import torch
